Address issue 18 from the clink-gizmos repo.
This adds a new fzf.allow_unsafe_query setting which allows the
fzf.exe_location setting to specify a file that doesn't have a ".exe"
extension.

On Windows, the `fzf --query` flag requires special quoting in order for
it to be safe.  But if the fzf.exe_location setting specifies a file
without a ".exe" extension, then the quoting may not work as expected,
since it may be intercepted and interpreted by CMD's batch script parser
or another scripting engine.  That could potentially interfere with the
effectiveness and safety of the quoting.

The fzf.lua script can compensate by stripping unsafe characters from
the query string before passing it to the non-exe wrapper script
specified in the fzf.exe_location setting.  But since that both reduces
functionality and also could potentially allow an unforeseen security
problem, a user must run `clink set fzf.allow_unsafe_query true` to give
permission to use that mode.

The `fzf --query` flag is only used by the `fzf_history` command, not by
`fzf_complete` or any of the other commands.